# services/places-service/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Building Management API
  description: API for managing locations, floors, and spaces (chill spaces, kitchens, meeting rooms, workspaces) in a building management system.
  version: v0
servers:
  - url: http://localhost:8080
    description: Local server
tags:
  - name: location-controller
    description: Operations for managing locations
  - name: floor-controller
    description: Operations for managing floors
  - name: space-controller
    description: Operations for managing spaces
paths:
  /api/v1/locations:
    get:
      tags:
        - location-controller
      operationId: getAllLocationsWithFloorsAndSpaces
      summary: Retrieve all locations, optionally including floors and spaces
      parameters:
        - name: includeFloors
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: includeSpaces
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LocationResponse'
    post:
      tags:
        - location-controller
      operationId: createLocation
      summary: Create a new location
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationResponse'
  /api/v1/locations/many:
    post:
      tags:
        - location-controller
      operationId: createLocations
      summary: Create multiple locations
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/LocationRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LocationResponse'
  /api/v1/locations/{id}:
    get:
      tags:
        - location-controller
      operationId: getLocationById
      summary: Retrieve a location by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationResponse'
    patch:
      tags:
        - location-controller
      operationId: updateLocation
      summary: Update a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationResponse'
    delete:
      tags:
        - location-controller
      operationId: deleteLocation
      summary: Delete a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  /api/v1/floors/{id}:
    get:
      tags:
        - floor-controller
      operationId: getFloorsByLocationId
      summary: Retrieve floors for a location, optionally including spaces
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: includeSpaces
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FloorResponse'
    post:
      tags:
        - floor-controller
      operationId: createFloorInLocationById
      summary: Create a new floor in a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FloorRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FloorResponse'
    patch:
      tags:
        - floor-controller
      operationId: updateFloor
      summary: Update a floor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FloorRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FloorResponse'
    delete:
      tags:
        - floor-controller
      operationId: deleteFloor
      summary: Delete a floor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  /api/v1/floors/many/{id}:
    post:
      tags:
        - floor-controller
      operationId: createFloorsInLocationById
      summary: Create multiple floors in a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/FloorRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FloorResponse'
  /api/v1/spaces/{id}:
    get:
      tags:
        - space-controller
      operationId: getSpacesByFloorId
      summary: Retrieve spaces for a floor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpaceResponse'
    post:
      tags:
        - space-controller
      operationId: createSpaceInFloorById
      summary: Create a new space in a floor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/ChillSpaceRequest'
                - $ref: '#/components/schemas/KitchenRequest'
                - $ref: '#/components/schemas/MeetingRoomRequest'
                - $ref: '#/components/schemas/WorkSpaceRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SpaceResponse'
    patch:
      tags:
        - space-controller
      operationId: updateSpace
      summary: Update a space
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/ChillSpaceRequest'
                - $ref: '#/components/schemas/KitchenRequest'
                - $ref: '#/components/schemas/MeetingRoomRequest'
                - $ref: '#/components/schemas/WorkSpaceRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SpaceResponse'
    delete:
      tags:
        - space-controller
      operationId: deleteSpace
      summary: Delete a space
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  /api/v1/spaces/many/{id}:
    post:
      tags:
        - space-controller
      operationId: createSpaceInFloorById_1
      summary: Create multiple spaces in a floor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                oneOf:
                  - $ref: '#/components/schemas/ChillSpaceRequest'
                  - $ref: '#/components/schemas/KitchenRequest'
                  - $ref: '#/components/schemas/MeetingRoomRequest'
                  - $ref: '#/components/schemas/WorkSpaceRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpaceResponse'
components:
  schemas:
    LocationRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 0
        street:
          type: string
          maxLength: 100
          minLength: 0
        buildingNumber:
          type: string
          maxLength: 10
          minLength: 0
        postalCode:
          type: string
          pattern: '\d{5}'
        phoneNumber:
          type: string
          pattern: '^\+?\d{10,15}$'
        email:
          type: string
          minLength: 1
        city:
          type: string
          maxLength: 50
          minLength: 0
        floors:
          type: array
          items:
            $ref: '#/components/schemas/FloorRequest'
      required:
        - name
        - street
        - buildingNumber
        - email
        - city
    LocationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        street:
          type: string
        buildingNumber:
          type: string
        postalCode:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        city:
          type: string
        floors:
          type: array
          items:
            $ref: '#/components/schemas/FloorResponse'
    FloorRequest:
      type: object
      properties:
        floorNumber:
          type: integer
          format: int32
        locationId:
          type: integer
          format: int64
        spaces:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/ChillSpaceRequest'
              - $ref: '#/components/schemas/KitchenRequest'
              - $ref: '#/components/schemas/MeetingRoomRequest'
              - $ref: '#/components/schemas/WorkSpaceRequest'
      required:
        - floorNumber
        - locationId
    FloorResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        floorNumber:
          type: integer
          format: int32
        locationId:
          type: integer
          format: int64
        spaces:
          type: array
          items:
            $ref: '#/components/schemas/SpaceResponse'
    SpaceRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 0
        seatingCapacity:
          type: integer
          format: int32
        floorId:
          type: integer
          format: int64
        spaceType:
          type: string
      required:
        - floorId
        - name
        - seatingCapacity
      discriminator:
        propertyName: spaceType
    SpaceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        seatingCapacity:
          type: integer
          format: int32
        floorId:
          type: integer
          format: int64
        spaceType:
          type: string
    ChillSpaceRequest:
      allOf:
        - $ref: '#/components/schemas/SpaceRequest'
        - type: object
          properties:
            numberOfSockets:
              type: integer
              format: int32
          required:
            - numberOfSockets
    KitchenRequest:
      allOf:
        - $ref: '#/components/schemas/SpaceRequest'
        - type: object
          properties:
            hasCoffeeMachine:
              type: boolean
            hasMicrowave:
              type: boolean
            hasFridge:
              type: boolean
          required:
            - hasCoffeeMachine
            - hasMicrowave
            - hasFridge
    MeetingRoomRequest:
      allOf:
        - $ref: '#/components/schemas/SpaceRequest'
        - type: object
          properties:
            hasProjector:
              type: boolean
            hasDrawingBoard:
              type: boolean
          required:
            - hasProjector
            - hasDrawingBoard
    WorkSpaceRequest:
      allOf:
        - $ref: '#/components/schemas/SpaceRequest'
        - type: object
          properties:
            hasComputer:
              type: boolean
            numberOfSockets:
              type: integer
              format: int32
            numberOfMonitors:
              type: integer
              format: int32
          required:
            - hasComputer
            - numberOfSockets
            - numberOfMonitors